casting as int

+++ payCalculator.py
@@ -4,7 +4,7 @@ def calculatePay():
     # This first line is provided for you
     hrs = input("Enter Hours:")
     rate = input('Enter Rate:')
-    gross_pay = int(hrs)*float(rate)
+    gross_pay = int(hrs)*int(rate)
     print(f'{gross_pay}')
     # end assignment
